fix(interbank): tolerate decimal public-stock amounts from peer banks
PublicStockSeller.Amount was int32; peer (Banka 2) sends fractional amounts (e.g. 9.0000 after a partial sale) → json decode of the whole list failed → remote public stocks silently empty. Changed to float64 (FE expects a number).

+++ services/bank-service/internal/domain/interbank.go
@@ -245,7 +245,9 @@ type TransactionVote struct {
 // PublicStockSeller — jedan prodavac (i njegova količina) za dati stock.
 type PublicStockSeller struct {
 	Seller ForeignBankId `json:"seller"`
-	Amount int32         `json:"amount"`
+	// Amount je float64 jer peer banke šalju decimalne količine (npr. 9.0000
+	// posle delimične prodaje); int32 bi obarao parsiranje cele liste.
+	Amount float64 `json:"amount"`
 }
 
 // PublicStock — agregirani entitet za GET /public-stock.
